Fix detection of options on findAndModify
There are two statements trying to establish the `options` object. The first does not consider the arguments length, but will set `options`. The 2nd is parametrized on `args.length` and will always return `false` since the last argument was `pop()`ed in the previous line.

This PR removes the redundant first assignment to `options` so that the 2nd can succeed.